* Implemented 3D reverb simulator plugin series - Room Builder Mono and Room Builder Stereo.
* Fixed improper InlineDisplay feature support in LV2 TTL files.
* Plugin names now have 'LSP' prefix for the LV2 format.
* Each plugin can now provide it's custom UI class which is derived from common plugin's UI class.
* Code now compiles for the PPC64 architecture.
* Code now compiles for the IBM s390x architecture.
* Added command-line option to pass configuration file name for standalone JACK plugins.
  The settings stored in configuration file will be automatically loaded at startup.
* Updated function signatures for libraries which now support '-fvisibility=hidden'
  compilation flag.
* Added support of nested menus in the UI toolkit.
* Added 3D scene rendering support in the UI by using different rendering backends.
* Implemented GLX backend based on openGL 2.x for rendering 3D scenes.
* Implemented key-value tree storage (KVTStorage) for storing and managing dynamic parameters.
* Implemented key-value tree storage (KVTStorage) UI<->DSP synchronization mechanisms.
* Implemented key-value tree storage (KVTStorage) serialization/deserialization in parameters.
* Implemented key-value tree storage (KVTStorage) serialization/deserialization for the plugin state.
* Implemented mechanism for UI<->DSP OSC message interchange.
* Added support of OSC protocol messages serialization, deserialization and pattern matching.
* Optimized complex number functions for AArch64 architecture.
* Changed installation path for jack core library from <lib-path> to <lib-path>/lsp-plugins.
* Added more careful file type analysis for non-EXT file systems.
* Added support of Hygon Dhyana x86 family CPU optimizations and detection of some other CPU vendors.
* Now both release and test binaries are available to build into separate subdirectories independently.
* Improved built-in resource generation tool.
* Embedded resources are now alphabetically sorted to make the build more deterministic.
* Added support of built-in presets for plugin's UI.

+++ include/dsp/arch/native/bswap.h
@@ -29,8 +29,8 @@ inline uint64_t __lsp_forced_inline     byte_swap(uint64_t v)
 inline float __lsp_forced_inline byte_swap(float v)
 {
     union {
-        float fValue;
-        uint32_t uValue;
+        float       fValue;
+        uint32_t    uValue;
     } tmp;
     tmp.fValue      = v;
     tmp.uValue      = byte_swap(tmp.uValue);
@@ -40,7 +40,7 @@ inline float __lsp_forced_inline byte_swap(float v)
 inline double __lsp_forced_inline byte_swap(double v)
 {
     union {
-        float fValue;
+        double   fValue;
         uint64_t uValue;
     } tmp;
     tmp.fValue      = v;
